+++ webui/src/router/index.js
@@ -19,6 +19,8 @@ import LanguageSelection from '@/views/LanguageSelection.vue'
 import LogLevelSelection from '@/views/LogLevelSelection.vue'
 import CpuGovernor from '@/views/CpuGovernor.vue'
 import DeviceMitigation from '@/views/DeviceMitigation.vue'
+import GpuGovernor from '@/views/GpuGovernor.vue'
+import GameTweaks from '@/views/GameTweaks.vue'
 
 window.wx = new WXEventHandler()
 
@@ -94,6 +96,16 @@ const routes = [
     path: '/settings/cpu_governor',
     name: 'CpuGovernor',
     component: CpuGovernor,
+  },
+  {
+    path: '/settings/gpu_governor',
+    name: 'GpuGovernor',
+    component: GpuGovernor,
+  },
+  {
+    path: '/settings/game_tweaks',
+    name: 'GameTweaks',
+    component: GameTweaks,
   },
   {
     path: '/settings/log_level',